feat: add REST API, status update and email consumers
- Add NotificationController: POST /api/notifications/send
- Add SendNotificationRequest DTO for REST input
- Add NotificationService.send() — save + route to producer
- Add NotificationStatusService: markAsSent, markAsFailed
- Add EmailConsumer: 3 @RabbitListener (HIGH/MEDIUM/LOW) with concurrency 5/3/1
- Extract processEmail(dto, priority) to avoid duplication
- Consumer updates status: SENT / FAILED with errorMessage
- Fix priorityType default: check null in toNotification
- Fix log.info — pass arguments instead of {}
- Remove TestController (temporary)

+++ src/main/java/zhulikov/project/notificationservice/repository/NotificationRepo.java
@@ -3,5 +3,5 @@ package zhulikov.project.notificationservice.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import zhulikov.project.notificationservice.entity.Notification;
 
-public interface NotificationRepo extends JpaRepository<Notification, Integer> {
+public interface NotificationRepo extends JpaRepository<Notification, Long> {
 }
